LayoutTests/http/tests/security/script-crossorigin-loads-correctly.html

<body>
<p>This test passes if the script loads correctly.</p>
<pre></pre>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
}

function done(msg) {
    document.querySelector("pre").innerHTML = msg;
    if (window.layoutTestController)
        layoutTestController.notifyDone();
}

var script = document.createElement("script");
script.crossOrigin = "";
script.src = "http://localhost:8000/security/resources/cors-script.php";
script.onload = function() { done("PASS"); }
script.onerror = function() { done("FAIL");}
document.body.appendChild(script);
</script>
