third_party/blink/web_tests/platform/generic/http/tests/security/script-crossorigin-loads-correctly-expected.txt

ALERT: script ran.
This test passes if the script loads correctly.

PASS
